# Conflicts:
#	DAL/Context/DatabaseContext.cs
#	DAL/DAL.csproj
#	DAL/Repository/Company/CompanyProjectRepository.cs
#	DAL/Repository/GenericCrudRepository.cs
#	DAL/Repository/Project/ProjectExperimentRepository.cs
#	DAL/Repository/Role/RolePermissionRepository.cs
#	DAL/Repository/User/UserCompanyRepository.cs
#	DAL/Repository/User/UserProjectRepository.cs
#	DAL/Repository/User/UserRoleRepository.cs
#	LOGIC/Services/Company/CompanyService.cs
#	LOGIC/Services/Experiment/ExperimentService.cs

+++ API/Controllers/Company/CompanyController.cs
@@ -1,4 +1,6 @@
+using System;
 using System.Web.Http;
+using DAL.Context;
 using Domain.Requests;
 using LOGIC.Services.Company;
 
@@ -8,17 +10,24 @@ namespace API.Controllers.Company
     public class CompanyController : ApiController
     {
         private readonly CompanyService _companyService;
-        
+
         public CompanyController()
         {
             _companyService = new CompanyService();
         }
-        
+
+        public CompanyController(IDatabaseContext context)
+        {
+            _companyService = new CompanyService(context);
+        }
+
         [HttpGet]
         [Route("companies")]
         public IHttpActionResult GetAll()
         {
-            var results = _companyService.GetAll(1, 1);
+            var results = _companyService.GetAll();
+            if (results.Count == 0) return Ok("There are no companies found");
+
             return Ok(results);
         }
 
@@ -27,30 +36,49 @@ namespace API.Controllers.Company
         public IHttpActionResult GetById(int id)
         {
             var result = _companyService.GetById(id);
+
             if (result == null) return NotFound();
             return Ok(result);
         }
 
         [HttpPost]
         [Route("companies")]
-        public IHttpActionResult Create(CompanyRequest entity)
+        public IHttpActionResult Create([FromBody] CompanyRequest companyToCreate)
         {
-            var result = _companyService.Create(entity);
+            if (companyToCreate == null) return BadRequest("Empty request body!");
+            if (companyToCreate.Name == null)
+                return BadRequest("You cannot create a company without specifying the name");
+            if (companyToCreate.Description == null)
+                return BadRequest("You cannot create a company without specifying the description");
+
+            var temp = _companyService.GetByName(companyToCreate.Name);
+
+            if (temp != null) return BadRequest("A company with that name already exists!");
+
+            var result = _companyService.Create(companyToCreate);
             return Ok(result);
         }
 
         [HttpPut]
         [Route("companies/{id}")]
-        public IHttpActionResult Update(int id, CompanyRequest entity)
+        public IHttpActionResult Update(int id, CompanyRequest companyRequest)
         {
-            var result = _companyService.Update(id, entity);
+            if (companyRequest == null) return BadRequest("Empty request body!");
+
+            var company = _companyService.GetById(id);
+            if (company == null) return NotFound();
+
+            var result = _companyService.Update(id, companyRequest);
             return Ok(result);
         }
 
+        //TODO:
         [HttpDelete]
         [Route("companies/{id}")]
         public IHttpActionResult Delete(int id)
         {
+            var company = _companyService.GetById(id);
+            if (company == null) return NotFound();
             var result = _companyService.Delete(id);
             return Ok(result);
         }
