+++ API/Controllers/User/UserController.cs
@@ -1,3 +1,6 @@
+using System;
+using System.Net.Mail;
+using System.Text.RegularExpressions;
 using System.Web.Http;
 using Domain.Requests;
 using LOGIC.Services.User;
@@ -8,7 +11,7 @@ namespace API.Controllers.User
     public class UserController : ApiController
     {
         private readonly UserService _userService;
-        
+
         public UserController()
         {
             _userService = new UserService();
@@ -18,7 +21,9 @@ namespace API.Controllers.User
         [Route("users")]
         public IHttpActionResult GetAll()
         {
-            var results = _userService.GetAll(1, 1);
+            var results = _userService.GetAll();
+            if (results.ToArray().Length == 0) return Ok("There are no users found");
+
             return Ok(results);
         }
 
@@ -33,17 +38,32 @@ namespace API.Controllers.User
 
         [HttpPost]
         [Route("users")]
-        public IHttpActionResult Create(UserRequest entity)
+        public IHttpActionResult Create([FromBody] UserRequest userToCreate)
         {
-            var result = _userService.Create(entity);
-            return Ok(result);
+            if (userToCreate == null) return BadRequest("Empty request body!");
+            if (userToCreate.Name == null) return BadRequest("You cannot create a user without specifying the name");
+            if (userToCreate.Email == null) return BadRequest("You cannot create a user without specifying the email");
+            if (!IsValid(userToCreate.Email)) return BadRequest("Invalid email address!");
+
+            var temp = _userService.GetByName(userToCreate.Name);
+            if (temp != null) return BadRequest("User with that name already exists!");
+            temp = _userService.GetByEmail(userToCreate.Email);
+            if (temp != null) return BadRequest("User with that email already exists!");
+
+            _userService.Create(userToCreate);
+            return Created("", userToCreate);
         }
 
         [HttpPut]
         [Route("users/{id}")]
-        public IHttpActionResult Update(int id, UserRequest entity)
+        public IHttpActionResult Update(int id, UserRequest userToUpdate)
         {
-            var result = _userService.Update(id, entity);
+            Console.WriteLine(userToUpdate.Email);
+            Console.WriteLine(userToUpdate.Name);
+            
+            var user = _userService.GetById(id);
+            if (user == null) return NotFound();
+            var result = _userService.Update(id, userToUpdate);
             return Ok(result);
         }
 
@@ -51,8 +71,19 @@ namespace API.Controllers.User
         [Route("users/{id}")]
         public IHttpActionResult Delete(int id)
         {
-            var result = _userService.Delete(id);
-            return Ok(result);
+            var result = _userService.GetById(id);
+            if (result == null) return NotFound();
+            _userService.Delete(id);
+            return Ok("User by id " + id + "has been deleted");
+        }
+
+        public bool IsValid(string emailAddress)
+        {
+            var emailRegex = @"^[\w!#$%&'*+\-/=?\^_`{|}~]+(\.[\w!#$%&'*+\-/=?\^_`{|}~]+)*"
+                             + "@"
+                             + @"((([\-\w]+\.)+[a-zA-Z]{2,4})|(([0-9]{1,3}\.){3}[0-9]{1,3}))$";
+
+            return Regex.IsMatch(emailAddress, emailRegex);
         }
     }
 }
